requirements.txt
telebot==0.0.5
telebot-calendar==1.2
python-dotenv==1.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
peewee==3.16.2
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-telegram-bot-calendar==1.0.5
requests==2.31.0
six==1.16.0
urllib3==2.0.2
